removing commit from git history
