+++ packages/valory/services/trader/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeigtuothskwyvrhfosps2bu6suauycolj67dpuxqvnicdrdu7yhtvq
 fingerprint_ignore_patterns: []
-agent: valory/trader:0.1.0:bafybeie6ejbn6rgf7nib7yd7qcskicv7ju4auhqgdgsiec3dd3jhsk26cu
+agent: valory/trader:0.1.0:bafybeici6n4h27owfdvno37jmaxlwyqungsg4w2tt7g4k7e6xk7yi6hudq
 number_of_agents: 4
 deployment:
   agent:
