+++ packages/valory/services/trader_pearl/service.yaml
@@ -8,7 +8,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeibg7bdqpioh4lmvknw3ygnllfku32oca4eq5pqtvdrdsgw6buko7e
 fingerprint_ignore_patterns: []
-agent: valory/trader:0.1.0:bafybeie6ejbn6rgf7nib7yd7qcskicv7ju4auhqgdgsiec3dd3jhsk26cu
+agent: valory/trader:0.1.0:bafybeici6n4h27owfdvno37jmaxlwyqungsg4w2tt7g4k7e6xk7yi6hudq
 number_of_agents: 1
 deployment:
   agent:
