+++ packages/valory/skills/decision_maker_abci/models.py
@@ -500,40 +500,6 @@ class DecisionMakerParams(MarketManagerParams, MechInteractParams):
         return Path(path)
 
 
-class BenchmarkingMode(Model, TypeCheckMixin):
-    """Configuration for the benchmarking mode."""
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        """Initialize the `BenchmarkingMode` object."""
-        self.enabled: bool = self._ensure("enabled", kwargs, bool)
-        self.native_balance: int = self._ensure("native_balance", kwargs, int)
-        self.collateral_balance: int = self._ensure("collateral_balance", kwargs, int)
-        self.mech_cost: int = self._ensure("mech_cost", kwargs, int)
-        self.pool_fee: int = self._ensure("pool_fee", kwargs, int)
-        self.sep: str = self._ensure("sep", kwargs, str)
-        self.dataset_filename: Path = Path(
-            self._ensure("dataset_filename", kwargs, str)
-        )
-        self.question_field: str = self._ensure("question_field", kwargs, str)
-        self.question_id_field: str = self._ensure("question_id_field", kwargs, str)
-        self.answer_field: str = self._ensure("answer_field", kwargs, str)
-        self.p_yes_field_part: str = self._ensure("p_yes_field_part", kwargs, str)
-        self.p_no_field_part: str = self._ensure("p_no_field_part", kwargs, str)
-        self.confidence_field_part: str = self._ensure(
-            "confidence_field_part", kwargs, str
-        )
-        # this is the mode for the p and confidence parts
-        # if the flag is `True`, then the field parts are used as prefixes, otherwise as suffixes
-        self.part_prefix_mode: bool = self._ensure("part_prefix_mode", kwargs, bool)
-        self.bet_amount_field: str = self._ensure("bet_amount_field", kwargs, str)
-        self.results_filename: Path = Path(
-            self._ensure("results_filename", kwargs, str)
-        )
-        self.randomness: str = self._ensure("randomness", kwargs, str)
-        self.nr_mech_calls: int = self._ensure("nr_mech_calls", kwargs, int)
-        super().__init__(*args, **kwargs)
-
-
 class AccuracyInfoFields(Model, TypeCheckMixin):
     """Configuration which holds the accuracy information file's fieldnames."""
