+++ packages/valory/skills/market_manager_abci/behaviours.py
@@ -39,7 +39,10 @@ from packages.valory.skills.market_manager_abci.graph_tooling.requests import (
     MAX_LOG_SIZE,
     QueryingBehaviour,
 )
-from packages.valory.skills.market_manager_abci.models import SharedState
+from packages.valory.skills.market_manager_abci.models import (
+    BenchmarkingMode,
+    SharedState,
+)
 from packages.valory.skills.market_manager_abci.payloads import UpdateBetsPayload
 from packages.valory.skills.market_manager_abci.rounds import (
     MarketManagerAbciApp,
@@ -68,6 +71,11 @@ class BetsManagerBehaviour(BaseBehaviour, ABC):
         """Get the shared state."""
         return cast(SharedState, self.context.state)
 
+    @property
+    def benchmarking_mode(self) -> BenchmarkingMode:
+        """Return the benchmarking mode configurations."""
+        return cast(BenchmarkingMode, self.context.benchmarking_mode)
+
     def store_bets(self) -> None:
         """Store the bets to the agent's data dir as JSON."""
         serialized = serialize_bets(self.bets)
@@ -91,7 +99,7 @@ class BetsManagerBehaviour(BaseBehaviour, ABC):
         """Read the bets from the agent's data dir as JSON."""
         self.bets = []
 
-        if self.shared_state.first_read:
+        if (not self.benchmarking_mode.enabled) and self.shared_state.first_read:
             # this is a temporary hack to overcome a multi-bets issue
             # if a bet that is in the `TO_PROCESS` queue cannot be selected because of the constraints
             # (e.g., not in opening margin), then everything is blocked because the `FRESH` status will never be updated:
