+++ packages/valory/skills/market_manager_abci/skill.yaml
@@ -8,7 +8,7 @@ aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   README.md: bafybeie6miwn67uin3bphukmf7qgiifh4xtm42i5v3nuyqxzxtehxsqvcq
   __init__.py: bafybeigrtedqzlq5mtql2ssjsdriw76ml3666m4e2c3fay6vmyzofl6v6e
-  behaviours.py: bafybeicutmw6uwldyrcr6vtcmdxioaja4n64cvzceccbeprgobw5mnei7a
+  behaviours.py: bafybeibvwoffr3cvfowjdlgb2mgbsjgac6xlthrf4f6kzjfwk3evi7qory
   bets.py: bafybeid3tzd3ikwdcyplltxttycrv3bixhxuxv56p3bfpru6gzigpoeu4i
   dialogues.py: bafybeiebofyykseqp3fmif36cqmmyf3k7d2zbocpl6t6wnlpv4szghrxbm
   fsm_specification.yaml: bafybeic5cvwfbiu5pywyp3h5s2elvu7jqdrcwayay7o3v3ow47vu2jw53q
@@ -22,7 +22,7 @@ fingerprint:
   graph_tooling/requests.py: bafybeibjyb6av33aswnptttekj6t7k7xysgphh2bigoorcgkc54y2j3xkm
   graph_tooling/utils.py: bafybeig5hxhnqgyfn5ym3poc5nziqwpeozqbd6wa4s6c2hjn6iyedg3t3y
   handlers.py: bafybeihot2i2yvfkz2gcowvt66wdu6tkjbmv7hsmc4jzt4reqeaiuphbtu
-  models.py: bafybeiei6kyz44z2juokg6moaadg54xnkaxbtqxm4swvrfcqbofhk73ali
+  models.py: bafybeicqsuhtzernjqwdiwy6gbfvtv4ashlxsfutnqldivxc5iufgk7sbu
   payloads.py: bafybeicfymvvtdpkcgmkvthfzmb7dqakepkzslqrz6rcs7nxkz7qq3mrzy
   rounds.py: bafybeibqqq3vjotaasc67olhlqthka6e6refodguntkmpksgdbqlzme73a
   tests/__init__.py: bafybeigaewntxawezvygss345kytjijo56bfwddjtfm6egzxfajsgojam4
@@ -151,6 +151,33 @@ models:
       the_graph_payment_required_error: payment required for subsequent requests for
         this API key
     class_name: MarketManagerParams
+  benchmarking_mode:
+    args:
+      enabled: false
+      native_balance: 10000000000000000000
+      collateral_balance: 10000000000000000000
+      mech_cost: 10000000000000000
+      pool_fee: 20000000000000000
+      outcome_token_amounts:
+      - 11000000000000000000
+      - 9000000000000000000
+      outcome_token_marginal_prices:
+      - 0.4
+      - 0.6
+      sep: ','
+      dataset_filename: benchmark_data.csv
+      question_field: question
+      question_id_field: question_id
+      answer_field: answer
+      p_yes_field_part: p_yes_
+      p_no_field_part: p_no_
+      confidence_field_part: confidence_
+      part_prefix_mode: true
+      bet_amount_field: collateral_amount
+      results_filename: benchmarking_results.csv
+      randomness: benchmarking_randomness
+      nr_mech_calls: 60
+    class_name: BenchmarkingMode
   network_subgraph:
     args:
       api_id: network
